Refactored the code
